GH-74: Рефактор RESTHandler с тестами

+++ core/src/main/java/ru/teamtwo/core/dtos/customer/CartItemDto.java
@@ -2,7 +2,7 @@ package ru.teamtwo.core.dtos.customer;
 
 public record CartItemDto(
     Integer id,
-    Integer customerId,
+    Long customerId,
     Integer productId,
     Integer quantity
 ){}
